+++ bop/src/value_conversion/error.rs
@@ -1,6 +1,6 @@
-#[cfg(feature = "no_std")]
+#[cfg(all(feature = "no_std", not(feature = "std")))]
 use alloc::{format, string::String, vec::Vec};
-#[cfg(not(feature = "no_std"))]
+#[cfg(any(feature = "std", not(feature = "no_std")))]
 use std::{format, string::String, vec::Vec};
 
 use core::fmt;